+++ contracts/finance/PaymentSplitter.sol
@@ -4,7 +4,6 @@ pragma solidity ^0.8.0;
 
 import "../utils/Address.sol";
 import "../utils/Context.sol";
-import "../utils/math/SafeMath.sol";
 
 /**
  * @title PaymentSplitter